test: Fix makeSequentialSimulationRequests tests
